UnmarkedTakbok (#7060)
## About The Pull Request
Fixes some spelling mistakes in relation to the Unmarked takbok.
Adds a recolored sprite for the weapon
Buffs the weapon some to keep it line with some nerfs around the weapon.
Still takes 4 shots to crit.
Added a hollowpoint ammo box to the takbok set.
## Why It's Good For The Game
Spelling is nice. Espically when they are real words.
Visual distinction is always appreciated
Buffs the damage and the fire time of the unmarked takbok. While nerfing
the unwielded recoil some. Hopefully to bring it back into being in
consideration of other weapons.
Ammo box is to act as a final buffer as the only way to my knowledge to
get the .585 ammo is via a ammo bench which the takbok does not come
with clips so I don't believe it can be refilled.
## Changelog
:cl:
add: Unmarked takbok has received a recolored sprite
add: Takbok kit comes with hollowpoint rounds.
balance: Unmarked takbok unwielded recoil has been nerfed (1.6 -> 2)
Time to fire reduced (0.8 -> 0.6) and damage multiplier 1 -> 1.3)
spellcheck: fixed a few typos in relation to the unmarked takbok
/:cl: