Budgeted meteor satellite crate (#6990)
## About The Pull Request
Added a new crate to the engineering department that only spawns meteor
satellites if they are not a current station goal.
## Why It's Good For The Game
Engineers can now protect the station more reliably without getting a
freebie goal completion.
## Changelog
:cl: EssentialTomato
add: Added a roundstart crate that only contains meteor satellites if
they are not a current station goal.
/:cl: