OOC pronouns on hover (#7036)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Adds pronouns that users can set in game preferences to display in OOC
when they speak.
Players can choose from a predefined list of pronouns. As much as I
would like to allow any pronouns, it's not a risk I'm willing to take
because someone will not use it for pronouns


![image](https://github.com/user-attachments/assets/ffc995c0-4bee-483b-813e-5af7ca5d15a7)

![image](https://github.com/user-attachments/assets/c9642bc1-f174-4994-83c2-ef5ddd49ab29)


## Why It's Good For The Game

Pronouns yes. QoL.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: OOC pronouns
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: Lucy <lucy@absolucy.moe>