Gives NT reps a good assistant points puncher (#7140)
## About The Pull Request

Gives the NT rep the ability to reward assistants with the silly little
corporate punchcard.

## Why It's Good For The Game

It makes sense from a flavour standpoint, it's corporate as hell and the
NT rep is the representative of the corporation. Hell, NT are the ones
regularly sending waves of interns at problems.

## Changelog


:cl:
add: Gives NT representatives a good assistant points punchcard.
/:cl: