Updates Icebox's service to be on par with TG (#7111)
## About The Pull Request

Basically the tin, this was something I had my eye on doing for a fair
while but only had the motivation to do so recently.

This is a rough port of:
https://github.com/tgstation/tgstation/pull/83163

As a side note, this PR removes iceboxes randomized bars until I have
time to fix them. In addition this does not add the minecart rails as I
will need assistance adding them in another PR.

I also made sure to remember the ranch, don't worry chicken maxxer bros.

![Screenshot 2025-07-06
025348](https://github.com/user-attachments/assets/e690ff82-0762-4b1f-8307-234882fa278b)


## Why It's Good For The Game
As the original PR states Bars should be front and center of the
station, this also gives the service department a higher amount of
cohesion. And it looks nicer. Ice is nice.

## Changelog
:cl: Thedragmeme
map: Updates Icebox's Service and connecting maint areas
/:cl:

---------

Co-authored-by: ThePooba <81843097+ThePooba@users.noreply.github.com>