Carved Square Pumpkin (Jackenstein Pumpkin) (#7205)
## About The Pull Request

Adds a "carved square pumpkin" hat to the game. This hat is based off of
Jackenstein from Deltarune Chapter 4. The hat would come as a free
edition to the loadout, categorized with the Halloween items (as it's a
carved pumpkin). Comes with a lighting toggle that while it does not
produce light, does set the state of the eyes to emissive and makes them
play a flickering light animation. The description also included a
direct reference to Chapter 4, including: "your taking too long".
## Why It's Good For The Game

While its uses are somewhat limited, it allows players to try out new
looks for their characters and express them better.

## Changelog
:cl:
add: Added a "carved square pumpkin" to the loadout selector under hats.
Note: The "toggle helmet light" does not actually produce laight.
/:cl: