[BOUNTY] Adds Jaundice (as a monkecoin purchase) (#7237)
## About The Pull Request
Adds a purchase to the roundstart shop that turns your skin yellow
## Why It's Good For The Game
It's funny, some people want it, and I believe it is ook approved?

Just hold off on merging for a little bit just for feedback purposes.
## Changelog

:cl:
add: Adds "Jaundice yourself" to the roundstart monkecoin shop, that
turns you skin yellow.
/:cl: