Changes assault operative's sunbeam alert level (#7441)
## About The Pull Request
Changes the alert level that is set when assault operatives get all of
their disks from gamma to delta
## Why It's Good For The Game
Gamma is meant to be an "event level" that admins set whenever they
wanna run some kind of event. Using it for assault operatives is not
only odd, considering it's supposed to "station destroying threat", but
the alert actively tells crew to stay in their departments (as if it was
just red). This also makes it consistent with other antags that destroy
the station, like nukies and blob.
## Changelog
:cl:
balance: Assault Operatives sunbeam will now set the alert level to
Delta
/:cl: