Deleting of diseases who live (#7517)
## About The Pull Request

- **REMOVES SENTIENT DISEASE**
## Why It's Good For The Game
Sentient disease is a relic of an older time at this point. And I figure
what better time than now to old yeller it from the codebase.
### Why?
A few reasons.

- Lack of interest.
We have several antagonists in codebase that arent being used. Example,
revolutionaries. Difference is, I don't see the same attachment or
interests in it. Most folks who do bring it up on discord is more a
relic of the past.

Key infection abilities, such as sneezing, are just nonfunctional, and
more would need to be changed to play with patho to be a somewhat
survivable antagonist.

- Removals elsewhere

To qoute their thoughts ["Not very fun to play, not fun to play against.
Not many folks are gonna miss it. This was a flat removal. Not much to
test here."](https://github.com/BeeStation/BeeStation-Hornet/pull/6472)

- Obligatory "Rats Rats we are the rats" joke here

Plague rats are our current spiritual successor to a bad virus that
messes up the station. And most of the time, folks can't even cure that
without the disease mutating.
## Changelog
:cl:
code: Nuked sentient disease code.
/:cl: