Reduces lone op starting/minimum weight (#7360)
## About The Pull Request
Reduces the lone op's starting and minimum weight to 0, down from 1
## Why It's Good For The Game
A lone op spawning on low pop is incredibly difficult to deal with, and
the original reason the minimum was increased to 1 was as a band-aid fix
to the antag not spawning. Now that their spawning has been fixed, this
band-aid solution is no longer needed, and low pop can rest easy (unless
they don't get the disk).
## Changelog
:cl:
balance: Lone Op's minimum/starting weight decreased to 0
/:cl: