Hawaii Expansion [BOUNTY] (#7563)
## About The Pull Request

Adds a bunch of pineapple - themed things from yogstation.
This includes new foods, a new drink sprite, a hula skirt and a new pet.

## Why It's Good For The Game
More content is fun, and this doesn't conflict with anything like bloat
content would. Plus ook wants this!
## Changelog
:cl:
add: Several new pineapple - themed food recipes
add: A new sprite for coconut milk in a glass
add: A hula skirt, that can be bought from autodrobes, or in the loadout
store
add: A new pet, the pineapple creacher, which is a smaller, living
pineapple.
fix: Coconuts now give coconut milk, and not... normal milk (why)
code: Adds a credits folder to the icons file for you to credit sprites.
Use it you dinguses.
/:cl:

---------

Co-authored-by: ThePooba <81843097+ThePooba@users.noreply.github.com>