Battlecruiser adjustment. (#7625)
## About The Pull Request
Removes the crew roles from the battlecruiser
## Why It's Good For The Game
The battlecruiser consists of 8 operatives, 4 crew, and the nuclear
leader.
Despite the objective of the 4 crew to not directly assault the station,
everyone gets giddy of "Oh battlecruiser, hell yeah". And joins in on
rushing the armory and the station.
So that being said removes the 4 crew from spawning and splits the 8
operatives amongst the two dorms.
## Changelog
:cl:
balance: Removed the 4 crew spawns from the battlecruiser starfury
/:cl: