Further edits to departmental bells (#7723)
## About The Pull Request
Adds in a syndicate version of the departmental bell, for mapping
purposes in the depot, or whatever. _i swear to god if you dont edit
bell_location on this-_
Updates heliostation to have dept. bells, somehow i forgot in my
original PR and noone noticed.
## Why It's Good For The Game
Might be neat for mapping, *shrug
Bug fix goob.
This threw me for a loop when i was originally coding because it was the
wrong solution but still got the right answer
## Changelog
:cl: Tractor Maam
map: heliostation now has departmental bells.
map: added a syndicate variant of the departmental bell for mapping
purposes.
del: Removed herobrine.
/:cl: