Adds silicon to job bonuses. (#7852)
## About The Pull Request
Adds department silicon to the monkecoin bonuses 
## Why It's Good For The Game
Silicons theoretically are both an important role and ongoing for the
station. Several systems and antagonist spawns with counting them as
enemies in mind, and the AI is the only way to be aware of probing and
hackings before they happen.
In addition, silicon's lack any base way to earn monkecoins itself. They
cant complete any stations objectives besides cargo budget and the BSA.
They can't do bounties. Ideally, we want people to play silicon's, so
give them a treat of the coins.
## Changelog
:cl:
add: Cyborgs and AI players now receive a end of our Monkecoin bonus
/:cl: