Makes the proto-kinetic railgun bulky instead of huge (#7838)
## About The Pull Request

this downgrades the pk railgun's size from huge to bulky, allowing it to
fit in the suit storage of things that allow for pkas in the suit
storage

## Why It's Good For The Game

the railgun being huge feels awkwardly arbitrary

the mining guns - which are far more powerful when used correctly - are
_normal sized_, and fit in a backpack along with their ammo, and as far
as I know, the railgun the only of the miner weapons that cannot fit in
suit storage of drake armor or whatever.

## Changelog
:cl:
balance: The proto-kinetic railgun is now bulky instead of huge,
allowing it to fit in the suit storage of certain mining suits/armors.
/:cl: