Lets you kick downed people in the groin for extra pain damage. (#7999)
## About The Pull Request
See title.
## Why It's Good For The Game
someone said this in OOC and it sounded funny
people you dont like getting kicked in the crotch is always funny
it has a glass cracking noise for extra comedic effect
## Changelog
:cl:
add: If you kick a downed person while aiming groin, hitting the chest
with your kick will deal 50 sharp brute pain to them. It also is
appropriately brutal. This effect is blocked if your target is wearing
thick clothing that covers the groin.
/:cl:

---------

Co-authored-by: ThrowawayUseless <notarealemail@emailservice.fake>