Removes the T2 tools from the CE's belt (#7964)
## About The Pull Request
Removes the T2 tools from CE belt
## Why It's Good For The Game
We should encourage heads of staff to pick the job because they want to
lead a department. Not because they want upgraded tools at roundstart.

I left in the RCD because its just a different sprite and an upgrade,
which can arguably be removed, but I dont think it's as bad as the T2
tools are.
## Changelog
:cl:
balance: Downgraded CE tools to T1 instead of T2
/:cl: