Deparadoxing Walter the bullterrier and Walter the Doberman (#8017)
## About The Pull Request
Seperates Walter the Doberman and Walter the Bullterrier
Sprite from what I find the be the source when walter was added, skyrat
[PR:8927](https://github.com/Skyrat-SS13/Skyrat-tg/pull/8927) But link
seems broken none the less
## Why It's Good For The Game
Okay back when blueshift was added from skyrat. Blueshields office was
added. (Before we even had a blueshield). Alongside this was Walter the
Doberman. (D. Walter) However D walter was given their raw subtype but
never a actual sprite. This means over time between confusion and
whatnot Walter D was replaced with Walter B as Walter D was just.
Invisible. This also lead to the slow removal of Walter B from lawyers
office which is a shame. This attempts to alleviate it.
It also helps with confusion on several maps as there would be several
walters in existance.
## Changelog
:cl:
add: Added sprite for Dobermans
map: Walter the Doberman and Walter the Bullterrier are now seperate on
the maps.
fix: You no longer will sometimes be killing goliath the miner pet for
kiling pet objectives
/:cl: