Hit the greys in their knees (#7879)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
### Gives some much-needed nerfs/balance to abductor

- Shrink rays now no longer strip, only forcing the affected to drop
everything in their hands
- Baton maxium charges increased to 5, from 3.
- Baton no longer applies a sleep for 2 minutes, but rather a paralysis
for 1 minute.
- Baton recharge rate increased to 15 seconds, from 5.
- Abductee status is removed when their gland is. Dont let your gland
get taken from you.
- Alien pistol has been fixed. [Porting PR 89949 from
TG](https://github.com/tgstation/tgstation/pull/89949), and will do
damage again.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
We can talk about balance all day and night but abductors are
single-handedly the worst example of current balance. And the fact we
have to have rules to keep them in check is a failing on their design
and not great gameplay.
### A conversion antag with no cure.
Currently, abductee status is incurable. Meaning you get forced into a
status with a often BS objective with no real cure or solution. And you
dont even get anything interestingly mechanically or buildup, your just
a glorfied brainwashing at the point of abduction.
I'm forwarding the idea of removing it alongside their gland, but I'm
also willing to hear other ideas such as lobotomy. This, however is
unacceptable from a game design standpoint.
### Shrink ray
I come at this from their rules perspective and a balance perspective,
Dropping literally everything at the drop of a dime is awful with little
to no counterplay than just... dodge I guess? At minimum, an assistant
has shoes, Id, pda, bag, jumpsuit, and a radio. And to reequip half of
that, you need to equip the jumpsuit first, specifically, which is in
the pile. Not to mention trying to find whatever weapon you wanted on
top of taking double damage now, and a slower movement speed. In a game
where controlling the pace of combat is what lets you win combat.

From rules wise, this shrink ray, when being used properly, causes
people to get taken for experiments, then promptly lose their things
because they can't reasonably be retrieved in time for the abduction.
Causing people to promptly loose key things because of course they
would. I've seen nuke ops happen because they shot the captain, and they
never got their bag back from the floor. And how would they know its
over in cargo? They dont remember it.

This is without mentioning or addressing its use as a breaching tool.
Swapping it from a full strip to a disarm + slow is way healthier than
it is currently.

### Why is all their power in their baton?
The baton is overtuned. Flat out
Addressing the actual numbers, a single abductor, with a single baton,
could keep 24 people sleeping permanently. Without losing charges over
time. Or escalating combat.
- Why is that?

One it has the most ridiculous stun numbers I've seen. A whopping
2-minute sleep. Which could be fine if it was expensive in the Baton's
economy. However. It's just not. It has never been. A baton charge is
worth 5 seconds of time. This is nothing. Double so when you remember
that one baton is all it takes to when a fight. Even more so when the
time to get a charge back is less than the duration of the effect.

So increase the baton's maximum charges so they can stun a few folks
initially. Then a slower but reasonable cooldown of 15 seconds, with a
reduced secondary stun time to make it so if you have more than 4 people
stunned, you should be going negative on your charges.

Next remove sleep as a status from the baton. [Mimicing this PR from
Iris station](https://github.com/IrisSS13/IrisStation/pull/563) Though
I'm upping the total numbers from 30 to 60 seconds, as mentioned before.
As they put it so, so, so, beautifully. ``` Because it's not fun to sit
afk for 5-10 minutes while waiting for sleep to wear off,```
And its not, you have no choice or options, not is it very interactive.
And this forces the abductors to actually use more of their kit! Namely,
the Alien silencer that goes frankly ignored.

### Finally their pistol.
Their pistol is both arguably weak and underused, but mostley as its
broken, 20 CELLULAR damage is not to be slept on.
Simply fixes what is there
# Is this perfect?
No.
God no.
But its frankly better than what we have now. And frankly they deserve
more nerfs.

And if Drifting contractors can survive on breadcrumbs compared to what
abductors have despite having the same kidnapping sthick? They can
manage with this nerf.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: Abductee status is now cured alongside the removal of the abductee
gland.
balance: Shrink rays no longer force drop all clothing, rather just
what's in the users hands.
balance: abductor batons now have 5 charges, recharge in 15 seconds,
from 3 charges and 5 seconds,
balance: Abductor batons now only induce paralysis for 1 minute, rather
than sleep for 2 minutes.
fix: Alien pistols will deal damage properly again
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->