Virus food is made through heating it now (#8205)
## About The Pull Request
Let me lay this out.
Originally, to make virus food, you had to cool it to 200k with water
and milk mixed in equal parts.
HOWEVER. water freezes at 274k. and below that, it becomes ice.
Ice and water are not interchangable, so ice cannot be used for the
virus food recipe, but virus food requires water at ice temperatures to
create. Effectively making this impossible to create.
## Why It's Good For The Game
I want to make virus food and not cry in the event the pathology
dispenser is inaccessible/destroyed
## Changelog
:cl: Tractor Maam
fix: Virus Food is no longer impossible to create. As a result, you have
to heat it to 400k now instead of cooling it to 200k.
/:cl: