Adds Construction Bags To Engivend (#8207)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

adds 5 construction bags to Engivend list

## Why It's Good For The Game

Engineering has a really big and tedious job. usually carrying around a
ton of tools and materials, yet for some reason these bags are relegated
to the lockers.

adding to the vend adds more to the supply to help engineers remember to
carry around apc components or tons of cable

but to note in the current state of engineering this is pretty
cumbersome, and should look to ways to help improve engineering
experience a bit

## Changelog


:cl:
add: Engi Vend vendors now stock with Construction bags

/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->