Adds TGMC brute and burn kits to DeForest (#8213)
## About The Pull Request
>"Hey so as it turns out, we tidied up our warehouses and found out we
had this cool product in stock. Turns out we've been >supplying the TGMC
with these kits since *ages ago*, I guess it wouldn't hurt to let you in
on this deal huh?"
>
>\- Unnamed DeForest Sales Rep

Adds the bruise and burn packs to DeForest, these replace any
sutures/meshes that they sell.
You can find these in frontier orders and print them in the frontier
wallmed.
Credits for TGMC for the sprites and the idea
## Why It's Good For The Game
I think it would be cool to reflavor DeForest to be the TGMC suppliers,
instead of just having the same stock that you find in the NT medkits.

These kits are strictly for healing brute and burn, they do nothing for
wounds. The upside though is that they stack up to 40 and will
automatically cycle through damaged limbs. They have less initial
healing, but treated limbs will passively regenerate until they are
healed (or take damage again) AKA how it works on TGMC


https://github.com/user-attachments/assets/870229f4-0883-4a68-855f-3b000dd7bd03

## Changelog
:cl:
add: DeForest now sells unique bruise and burn packs
/:cl: