Makes the lava staff useful for terraforming lavaland (#8230)
## About The Pull Request

this makes it so that the lava staff's cooldowns are reduced on
lavaland: 10 -> 5 seconds to create lava, 5 -> 0.5 seconds to remove
lava.

cooldowns outside of lavaland are unaffected.

also cleaned up the code a bit.

## Why It's Good For The Game

usually the lava staff is really only used to like, put lava underneath
doors and such, so this gives it more of a proper _utility_ use too.

## Changelog
:cl:
balance: The cooldowns for using the lava staff are now heavily reduce
on lavaland, making it far more viable to use the staff to terraform
lavaland with it.
/:cl:

---------

Co-authored-by: Xander3359 <66163761+Xander3359@users.noreply.github.com>