pen fixes + blood sucker megafix (#8626)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Two bugs for pens fixed, almost every issue for blood sucker fixed.
Closes #7742, #5983, #6299, #7405, #7972,#8142,

I'm pretty sure #8090 is due to the player reconnecting, and should be
closed on it's own. #8618 is not a bug. #7753 I'm honestly not sure
about, but probably needs to be looked into. Finally, absolucy was
planning to look into #8259. And that's ALL the issues I could find.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Bugs bad. I'm not going through every individual fix.

fix: #7742
fix: #5983
fix: #6299
fix: #7405
fix: #7972 
fix: #8142 
fix: #8635 
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
fix: Bug-suckers have had an insane amount of bugs squashed. I mean,
seriously, these guys had a tracker dedicated to how buggy they were.
fix: Changing the pen angle is now on control click, instead of being on
Z (conflicted with clicking the pen)
fix: Clicking a pen no longer makes it large, and instead keeps it tiny.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->