Oozelings now properly lose blood when starving (#8698)
## About The Pull Request

This makes it so oozelings lose blood when starving.

I believe this is fine during the feature freeze, as it seems **they
were already meant to do this**, as one of their species menu
descriptions says this:

> Once **hungry enough**, Oozelings will begin to consume their own
blood and limbs.

and this was one of the things discussed about oozeling balance before
the feature freeze was put in place.

Also slightly reorganized the code

## Why It's Good For The Game

things working as described is nice, and this does help with giving them
a disadvantage to further balance them.

## Changelog
:cl:
balance: Oozelings will now begin to lose blood if they're starving, as
intended.
/:cl: