remove most restrictions on what areas a bloodsucker can use as a lair (#8838)
## About The Pull Request

this allows bloodsuckers to claim most areas as lairs, with only the
following areas disallowed:
- shuttles
- the centcom z-level
- outdoors areas and ruins (you can still make a new area on
lavaland/icebox/etc, tho)

## Why It's Good For The Game

want a lavaland base? sure. you'll have to deal with the effort of
bringing people there or going back to the station, but if you wanna,
then by all means!

## Changelog
:cl:
qol: Removed most restrictions on areas that can be used as a
Bloodsucker lair. Outdoor areas and shuttles are notable exceptions.
/:cl: