Update PULL_REQUEST_TEMPLATE.md (#8959)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Changes the PR template to include testing and a pre-merge checklist,
from Eris and Vanderlin.

## Why It's Good For The Game

Getting real tired of some people not testing their PRs. And if you
think we don't need it, without diligence, quality slacks. <!-- and to
behonest im going to fucking crash out if we don't do something like
this that works towards keeping the codebase **intact** -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
code: changed the pull request templating to include testing, and
pre-merge checklist
/:cl: